EES-4770 Move GetReleaseSeriesView into PublicationService
