+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Controllers/Api/PublicationController.cs
@@ -131,6 +131,14 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Controllers.Api
                 .HandleFailuresOrOk();
         }
 
+        [HttpGet("publications/{publicationId:guid}/release-series-view")]
+        public async Task<ActionResult<List<ReleaseSeriesItemViewModel>>> GetReleaseSeriesView(Guid publicationId)
+        {
+            return await _publicationService
+                .GetReleaseSeriesView(publicationId)
+                .HandleFailuresOrOk();
+        }
+
         [HttpPatch("api/publications/{publicationId:guid}/release-series-view")]
         public async Task<ActionResult<List<ReleaseSeriesItemUpdateRequest>>> UpdateReleaseSeries(
             Guid publicationId,
