+++ src/explore-education-statistics-admin/src/pages/publication/PublicationLegacyReleasesPage.tsx
@@ -1,11 +1,10 @@
 import LegacyReleasesTable from '@admin/pages/publication/components/LegacyReleasesTable';
 import usePublicationContext from '@admin/pages/publication/contexts/PublicationContext';
-import legacyReleaseService, {
-  ReleaseSeriesItem,
-} from '@admin/services/legacyReleaseService';
+import { ReleaseSeriesItem } from '@admin/services/legacyReleaseService';
 import LoadingSpinner from '@common/components/LoadingSpinner';
 import useAsyncHandledRetry from '@common/hooks/useAsyncHandledRetry';
 import React from 'react';
+import publicationService from '@admin/services/publicationService';
 
 const PublicationLegacyReleasesPage = () => {
   const { publicationId, publication } = usePublicationContext();
@@ -13,7 +12,7 @@ const PublicationLegacyReleasesPage = () => {
   const { value: releaseSeries = [], isLoading } = useAsyncHandledRetry<
     ReleaseSeriesItem[]
   >(
-    async () => legacyReleaseService.getReleaseSeriesView(publicationId),
+    async () => publicationService.getReleaseSeriesView(publicationId),
     [publicationId],
   );
 
@@ -24,9 +23,9 @@ const PublicationLegacyReleasesPage = () => {
   return (
     <>
       <h2>Legacy releases</h2>
-      <LegacyReleasesTable // @MarkFix rename?
+      <LegacyReleasesTable // @MarkFix rename to ReleaseSeriesTable?
         canManageLegacyReleases={
-          publication.permissions.canManageLegacyReleases // @MarkFix rename?
+          publication.permissions.canManageLegacyReleases // @MarkFix rename to canManageReleaseSeries?
         }
         releaseSeries={releaseSeries}
         publicationId={publicationId}
